--- Helpers/ShuffleHelper.cs
@@ -1,24 +1,0 @@
-﻿namespace TOHE.Modules;
-
-public static class ShuffleListExtension
-{
-    /// <summary>
-    /// Shuffles all elements in a collection randomly
-    /// </summary>
-    /// <typeparam name="T">The type of the collection</typeparam>
-    /// <param name="collection">The collection to be shuffled</param>
-    /// <param name="random">An instance of a randomizer algorithm</param>
-    /// <returns>The shuffled collection</returns>
-    public static IEnumerable<T> Shuffle<T>(this IEnumerable<T> collection, IRandom random)
-    {
-        var list = collection.ToList();
-        int n = list.Count;
-        while (n > 1)
-        {
-            n--;
-            int k = random.Next(n + 1);
-            (list[n], list[k]) = (list[k], list[n]);
-        }
-        return list;
-    }
-}
